move util array functions to submodule
